cmod_a7 support: Transition to amaranth_boards.
